i18n: fix redundant domain prefixes, convert hardcoded hints, and add bundled resource path for Windows/Linux
Root cause: tr!("settings", "settings-ai-...") produced lookup key
"settings-settings-ai-..." which never matched the FTL entries.
Same issue for "billing", "code" domains.

Changes:
- Remove redundant domain prefix from 417 tr!/static_tr! calls
  across 13 files (settings: 74+182, billing: 161, code: 2)
- Convert 6 hardcoded English strings in terminal/input.rs to
  tr!/static_tr! with corresponding FTL entries:
  - 20 rotating "Warp anything e.g." hints → static_tr! with
    existing agent-hint-* FTL keys
  - 2 queue hint constants → static_tr!
  - 3 child-agent format strings → tr!() with agent_name interpolation
  - 2 error toast strings → tr!()
- Add Windows exe-relative and Linux /opt path resolution in
  bundle_resources_dir() for packaged builds
- Add i18n directory copy to prepare_bundled_resources.ps1
- Add new FTL entries (en-US + zh-CN): steer-child-agent,
  queue-child-follow-up, ask-child-follow-up,
  images-removed-model-unsupported, conversation-navigate-error

+++ app/src/ai_assistant/requests.rs
@@ -400,11 +400,11 @@ impl Requests {
                 let num_hours = num_minutes / 60;
                 let num_days = num_hours / 24;
                 let remaining_text = if num_days > 0 {
-                    crate::tr!("ai_assistant", "duration-days", count = num_days as usize)
+                    crate::tr!("ai_assistant", "duration-days", count = num_days as i64)
                 } else if num_hours > 0 {
-                    crate::tr!("ai_assistant", "duration-hours", count = num_hours as usize)
+                    crate::tr!("ai_assistant", "duration-hours", count = num_hours as i64)
                 } else {
-                    crate::tr!("ai_assistant", "duration-minutes", count = num_minutes as usize)
+                    crate::tr!("ai_assistant", "duration-minutes", count = num_minutes as i64)
                 };
                 Some(remaining_text)
             }
